% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{andromedaScorePlot}
\alias{andromedaScorePlot}
\title{Plot Andromeda scores}
\usage{
andromedaScorePlot(evidence, pattern = NULL)
}
\arguments{
\item{evidence}{Output of read_evidence()}

\item{pattern}{Regex pattern of any experimental columns to exclude}
}
\value{
ggplot object
}
\description{
Plot Andromeda scores
}
