% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cor_plot}
\alias{cor_plot}
\title{Plot heatmap of Pearson Correlation}
\usage{
cor_plot(normDat, quantification = c("LFQ", "SILAC"))
}
\arguments{
\item{normDat}{output of either normaliseData()}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantification method was used for experiment}
}
\value{
ggplot2 heatmap
}
\description{
Plot heatmap of Pearson Correlation
}
