% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{count_plot}
\alias{count_plot}
\title{Plot no. quantified}
\usage{
count_plot(
  normDat,
  quantification = c("LFQ", "SILAC"),
  rep_pattern = "_0?[1-9]$"
)
}
\arguments{
\item{normDat}{Output of either normalise_sty() or normalise_proteinGroups()}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantification method was used for experiment}

\item{rep_pattern}{Regex string to group experimental conditions
\itemize{
\item i.e. which string defines the replicates
}}
}
\value{
ggplot2 object
}
\description{
Plot no. quantified
}
