% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extract_experiment_cols}
\alias{extract_experiment_cols}
\title{Extract columns containing quantitative data}
\usage{
extract_experiment_cols(
  dat,
  quantification = c("LFQ", "SILAC"),
  pattern = NULL
)
}
\arguments{
\item{dat}{MQ data frame}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantificiation method was used for experiment}

\item{pattern}{Regex string to Subset of experimental columns,
e.g. just proteome or STY columns. Default = NULL.}
}
\value{
vector of  columns containing experimental values
}
\description{
Extract columns containing quantitative data
}
