% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_proteinGroups}
\alias{filter_proteinGroups}
\title{Filter MaxQuant proteinGroups.txt}
\usage{
filter_proteinGroups(proteinGroups, quantification = c("LFQ", "SILAC"))
}
\arguments{
\item{proteinGroups}{output of read_proteinGroups}

\item{quantification}{one of LFQ or SILAC, depending on type
of quantification used in experiment.}
}
\value{
filtered proteinGroups dataframe
}
\description{
Filter MaxQuant proteinGroups.txt
}
