% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{filter_sty}
\alias{filter_sty}
\title{Filter MaxQuant Phospho (STY).txt}
\usage{
filter_sty(sty, quantification = c("LFQ", "SILAC"))
}
\arguments{
\item{sty}{output of read_phosphoSTY}

\item{quantification}{one of LFQ or SILAC, depending on type
of quantification used in experiment.}
}
\value{
filtered STY dataframe
}
\description{
Filter MaxQuant Phospho (STY).txt
}
