% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{imputeTruncNorm}
\alias{imputeTruncNorm}
\title{Impute Data by drawing randomly from a distribution}
\usage{
imputeTruncNorm(filteredData)
}
\arguments{
\item{filteredData}{output of filter_sty() or filter_proteinGroups()}
}
\value{
imputed values
}
\description{
The default represents a stimulation to represent a
population below the limits of detection, as described in
Tyanova et al., 2016
ONLY suitable for label free data!!
}
