% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{massErrorDensity}
\alias{massErrorDensity}
\title{Plot Mass Error Density}
\usage{
massErrorDensity(evidence, pattern = NULL)
}
\arguments{
\item{evidence}{Output of read_evidence()}

\item{pattern}{Regex pattern of any experimental columns to exclude}
}
\value{
ggplot object
}
\description{
Plot Mass Error Density
}
