% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{modificationCharacteristics}
\alias{modificationCharacteristics}
\title{Pie charts of modification characteristics}
\usage{
modificationCharacteristics(
  raw_modTable,
  modsp,
  quantification = c("LFQ", "SILAC")
)
}
\arguments{
\item{raw_modTable}{Output of read_phosphoSTY() or similar}

\item{modsp}{Output of read_modSpPeptides()}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantification method was used for experiment}
}
\value{
three ggplot objects as list
}
\description{
Distribution of characteristics of modifified
peptides, e.g. modified residue, localisation probability
}
