% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{normaliseData}
\alias{normaliseData}
\title{Normalise quantitative values}
\usage{
normaliseData(filteredData, quantification = c("LFQ", "SILAC"))
}
\arguments{
\item{filteredData}{output of filter_sty() or filter_proteinGroups()}

\item{quantification}{one of LFQ or SILAC, depending on type
of quantification used in experiment.}
}
\value{
Normalised dataframe
}
\description{
Normalise quantitative values
}
