% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_plot}
\alias{pca_plot}
\title{Principal component analysis}
\usage{
pca_plot(
  normDat,
  quantification = c("LFQ", "SILAC"),
  rep_pattern = "_0?[1-9]$",
  colour_grp_pattern = NULL,
  shape_grp_pattern = NULL,
  colour_scale = NULL
)
}
\arguments{
\item{normDat}{Output of either normalise_sty() or normalise_proteinGroups()}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantification method was used for experiment}

\item{rep_pattern}{Regex string to group replicates of same
experimental conditions}

\item{colour_grp_pattern}{Regex string to group experimental conditions
\itemize{
\item i.e. same stimulant, same time point
and projected as colour aesthetic in plot
}}

\item{shape_grp_pattern}{Regex string to group experimental conditions
i.e. same stimulant, same time point
and projected as shape aesthetic in plot.}

\item{colour_scale}{vector (can be named based on variables names in plot) of
colours}
}
\value{
ggplot2 object
}
\description{
Principal component analysis
}
