% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{phosphoenrichment}
\alias{phosphoenrichment}
\title{Calculate enrichment efficiency}
\usage{
phosphoenrichment(modsp, quantification = c("LFQ", "SILAC"), pattern = NULL)
}
\arguments{
\item{modsp}{Output of read_modSpPeptides()}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantificiation method was used for experiment}

\item{pattern}{Regex string to Subset of experimental columns,
e.g. just proteome or STY columns. Default = NULL.}
}
\value{
numeric value, percentage peptides with STY site identified
}
\description{
Calculate enrichment efficiency
}
