% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phosphoSTYSites}
\alias{phosphoSTYSites}
\title{Example STY MaxQuant output table that can be processed using this package.}
\format{
Data frames, as generated by MaxQuant. Details of output tables can
be found at:
\url{http://www.coxdocs.org/doku.php?id=maxquant:table:sitetable}
}
\source{
Internal (CFR lab) mass spec experiment, processed by MaxQuant
}
\usage{
phosphoSTYSites
}
\description{
Example STY MaxQuant output table that can be processed using this package.
}
\keyword{datasets}
