% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotPie}
\alias{plotPie}
\title{plot pie charts for modification charactersitics}
\usage{
plotPie(dat)
}
\arguments{
\item{dat}{2 column data frame containing name of pie slices
and n of each}
}
\value{
ggplot2 object
}
\description{
plot pie charts for modification charactersitics
}
