% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{read_experiments_names}
\alias{read_experiments_names}
\title{Import MQ output files}
\usage{
read_experiments_names(folderPath)
}
\arguments{
\item{folderPath}{Path of MQ txt output files}
}
\description{
Import MQ output files
}
