% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{read_proteinGroups}
\alias{read_proteinGroups}
\alias{read_phosphoSTY}
\alias{read_modSpPeptides}
\alias{read_peptides}
\alias{read_evidence}
\title{Import MQ output files}
\usage{
read_proteinGroups(
  folderPath,
  pattern = NULL,
  quantification = c("LFQ", "SILAC"),
  SILAC_ratios = c("all", "H/L and M/L", "H/L", "M/L")
)

read_phosphoSTY(
  folderPath,
  pattern = NULL,
  quantification = c("LFQ", "SILAC"),
  SILAC_ratios = c("all", "H/L and M/L", "H/L", "M/L")
)

read_modSpPeptides(
  folderPath,
  quantification = c("LFQ", "SILAC"),
  SILAC_ratios = c("all", "H/L and M/L", "H/L", "M/L")
)

read_peptides(folderPath)

read_evidence(folderPath)
}
\arguments{
\item{folderPath}{Path of MQ txt output files}

\item{pattern}{Regex string to Subset of experimental columns,
e.g. just proteome or STY columns. Default = NULL.}

\item{quantification}{One of "LFQ" or "SILAC", depending
which quantificiation method was used for experiment}

\item{SILAC_ratios}{If quantification is SILAC, which ratios to include?
Defaults to "all".}
}
\value{
data in format compatible for further analysis
}
\description{
Import MQ output files
}
